Weapon inventory and bug fixes
