+++ scripts/entities/items/weapons/weapon.py
@@ -665,18 +665,18 @@ class Weapon(Item):
         self.rendered_image.blit(dark_surface, (0, 0), special_flags=pygame.BLEND_RGBA_MULT)
 
 
-    # Inventory Logic below
-    #######################################################
-    # Pick up the weapon and update the general light in the area
-    def Pick_Up(self):
-        if self.in_inventory:
-            return False
-        self.entity = super().Pick_Up()
+    # # Inventory Logic below
+    # #######################################################
+    # # Pick up the weapon and update the general light in the area
+    # def Pick_Up(self):
+    #     if self.in_inventory:
+    #         return False
+    #     self.entity = super().Pick_Up()
 
-        if not self.entity:
-            return False
-        self.Pickup_Reset_Weapon(self.entity)
-        return True
+    #     if not self.entity:
+    #         return False
+    #     self.Pickup_Reset_Weapon(self.entity)
+    #     return True
     
     def Pickup_Reset_Weapon(self, entity):
         self.entity = entity
@@ -686,145 +686,6 @@ class Weapon(Item):
         self.enemy_hit = False
         self.light_level = 10
 
-    def Set_Equipped_Position(self, direction_y):
-        if 'left' in self.inventory_type:
-            if direction_y < 0:
-                self.Move((self.entity.pos[0] - 5 , self.entity.pos[1]))
-            else:
-                self.Move((self.entity.pos[0] + 5 , self.entity.pos[1]))
-        elif 'right' in self.inventory_type:
-            if  direction_y < 0:
-                self.Move((self.entity.pos[0] + 7, self.entity.pos[1]))
-            else:
-                self.Move((self.entity.pos[0] - 7, self.entity.pos[1]))
-        else:
-            print("DIRECTION NOT FOUND", self.inventory_type)
-
-
-    # Initialise the double clikc
-    def Handle_Double_Click(self, sending_inventory, receiving_inventory):
-        # Check if there is a free inventory slot
-        recieving_inventory_slot = None
-        for i in range(8):
-            recieving_inventory_slot = receiving_inventory.Find_Available_Inventory_Slot(recieving_inventory_slot)
-            if not self.Check_Legal_Move(recieving_inventory_slot, sending_inventory, receiving_inventory):
-                recieving_inventory_slot = None
-                continue
-
-            break
-            
-        
-        if not recieving_inventory_slot:
-            self.Reset_Inventory_Slot(sending_inventory)
-            return False
-        # Check if we can send the item to the new inventroy slot
-        if self.Send_To_Inventory(recieving_inventory_slot, sending_inventory, receiving_inventory):
-            return True
-        else:
-            self.Reset_Inventory_Slot(sending_inventory)
-            return False
-
-    def Check_Legal_Move(self, inventory_slot, sending_inventory, receiving_inventory):
-        if not inventory_slot:
-            return False
-        
-        if not self.Check_Two_Handed(inventory_slot, sending_inventory, receiving_inventory):
-            return False
-    
-        if not self.Check_For_Two_Handed_In_Weapon_Inventory(inventory_slot, sending_inventory, receiving_inventory):
-            return False
-        
-        if not self.Check_Two_Handed_Left_Hand(inventory_slot):
-            return False
-        
-        if not self.Bow_Check(inventory_slot):
-            return False
-        
-        if not self.Arrow_Check(inventory_slot):
-            return False
-        
-        return True
-
-
-    # Attempt to move the item to the receiving inventory slot by double clicking
-    def Send_To_Inventory(self, inventory_slot, sending_inventory, receiving_inventory):
-       
-        if not self.Check_Legal_Move(inventory_slot, sending_inventory, receiving_inventory):
-            return False
-
-        # Move the item
-        move_successful = receiving_inventory.Move_Item(self, inventory_slot)
-        # If the move was successful, remove it from the sending inventory
-        if move_successful:
-            self.Reset_Weapon()
-            # Remove item from old inventory and save the inventory type
-            inventory_type_holder = self.inventory_type
-
-            sending_inventory.Remove_Item(self, True)
-            # send weapon into weapon inventory, checked by seeing if it has an inventory_type
-            if self.inventory_type:
-
-                # If weapon is already equipped in the other hand, remove it before adding to new hand
-                if self.equipped:
-                    self.game.player.Remove_Active_Weapon(inventory_type_holder)
-                self.Equip()
-            else: # Drag weapon back into item inventory
-                self.Set_Equip(False)
-                self.game.player.Remove_Active_Weapon(inventory_type_holder)
-
-            return True
-        
-        
-        return False
-    
-    def Reset_Weapon(self):
-        self.rotate = 0
-        self.attacking = 0
-        self.sub_type = self.type
-        self.animation = 0
-
-    # Add weapons from a sending inventory to a receiving inventory using drag
-    # Example: weapon_inventory (sending) -> inventory (receiving) 
-    def Move_To_Other_Inventory(self, sending_inventory, receiving_inventory, offset = (0,0)):
-        if self.game.mouse.left_click:
-            return
-        # Check for collision with new ivnentory slot
-        for receiving_inventory_slot in receiving_inventory:
-            if receiving_inventory_slot.rect().colliderect(self.game.mouse.rect_pos(offset)):
-                if self.Send_To_Inventory(receiving_inventory_slot, sending_inventory, receiving_inventory):
-                    for inventory_slot in sending_inventory.inventory:
-                        if not inventory_slot.active: # Reset active state after dragging
-                            continue
-                        inventory_slot.Set_Active(False)
-                        break
-                    return True             
-        return False
-
-    # Check if the weapon can be moved to the weapon inventory
-    def Move_Inventory_Check(self, offset = (0,0)):
-        if not self.picked_up:
-            active_inventory = self.game.weapon_inventory.active_inventory
-            weapon_inventory = self.game.weapon_inventory.inventories[active_inventory]
-            if self.equipped: # Move to normal inventory
-                if self.Move_To_Other_Inventory(weapon_inventory, self.game.item_inventory, offset):
-                    self.Set_Equip(False)
-                    self.game.player.Remove_Active_Weapon(self.inventory_type)
-                    return True
-                
-                
-            else: # Move to weapon inventory
-                if self.Move_To_Other_Inventory(self.game.item_inventory, weapon_inventory, offset):
-                    self.Equip()
-                    return True
-                
-        return False
-    
-    # Equip the weapon
-    def Equip(self):
-        self.Set_Equip(True)
-        self.Set_Entity(self.game.player)
-        self.game.player.Set_Active_Weapon(self, self.inventory_type)
-
     def Set_Equip(self, state):
         self.equipped = state
         if state:
@@ -832,134 +693,269 @@ class Weapon(Item):
         else:
             self.render = True
 
-
-    # Check for out of bounds, return true if valid, else false
-    def Move_Legal(self, mouse_pos, player_pos, tilemap, offset = (0,0)):
-
-        # Check if the weapon can be moved to the weapon inventory
-        if self.Move_Inventory_Check(offset):
-            self.picked_up = True
-            self.move_inventory_slot = True
-            return False
-        if super().Move_Legal(mouse_pos, player_pos, tilemap, offset):
-            return True
+    def Set_Equipped_Position(self, direction_y):
+        if not self.entity:
+            return
+        if direction_y < 0:
+            self.Move((self.entity.pos[0] - 5 , self.entity.pos[1]))
         else:
-            return False
+            self.Move((self.entity.pos[0] + 5 , self.entity.pos[1]))
         
-    def Check_Inventory_Type(self, type):
-        legal_inventories = ['left_hand', 'right_hand', 'bow', 'arrow']
-        if type not in legal_inventories:
-            return False
-        
-        self.inventory_type = type
-        return True
 
-    def Update_Player_Hand(self, prev_hand):
-        # Check if the weapon has changed hands
-        if self.inventory_type != prev_hand:
-            self.Set_Equip(True)
-            if prev_hand == 'left_hand':
-                self.game.player.Remove_Active_Weapon(prev_hand)
-                self.game.player.Set_Active_Weapon(self, self.inventory_type)
-            elif prev_hand == 'right_hand':
-                self.game.player.Remove_Active_Weapon(prev_hand)
-                self.game.player.Set_Active_Weapon(self, self.inventory_type)
-            elif 'bow' in prev_hand:
-                self.game.player.Remove_Active_Weapon(prev_hand)
-                self.game.player.Set_Active_Weapon(self, self.inventory_type)
+    # # Initialise the double clikc
+    # def Handle_Double_Click(self, sending_inventory, receiving_inventory):
+    #     # Check if there is a free inventory slot
+    #     recieving_inventory_slot = None
+    #     for i in range(8):
+    #         recieving_inventory_slot = receiving_inventory.Find_Available_Inventory_Slot(recieving_inventory_slot)
+    #         if not self.Check_Legal_Move(recieving_inventory_slot, sending_inventory, receiving_inventory):
+    #             recieving_inventory_slot = None
+    #             continue
 
-
-    def Place_Down(self):
-        if not super().Place_Down():
-            return False
-        self.entity = None
-        self.in_inventory = False
-        self.picked_up = False
-        if self.equipped:
-            self.game.player.Remove_Active_Weapon(self.inventory_type)
-            self.Set_Equip(False)
-        return True
-
-    def Set_In_Inventory(self, state):
-        self.in_inventory = state
-
-    def Bow_Check(self, inventory_slot):
-        if not inventory_slot.inventory_type:
-            return True
-        if 'bow' in inventory_slot.inventory_type:
-            if 'bow' in self.type:
-                return True
-            else:
-                return False
-        return True
-        
-        
-    def Arrow_Check(self, inventory_slot):
-
-        if not inventory_slot.inventory_type:
-            return True
-        if 'arrow' in inventory_slot.inventory_type:
-            if 'arrow' in self.weapon_class:
-                return True
-            else:
-                return False
-        return True
-
-    def Check_Two_Handed(self, inventory_slot, sending_inventory, receiving_inventory):
-        if not 'two' in self.weapon_class:
-            return True
-        if inventory_slot.inventory_type:
-            # Try to find the inventory_slot, only the weapon inventory has this property
-            try:
-                if receiving_inventory.Find_Inventory_Slot(inventory_slot):
-                    # Find the original position of the item in the inventory
-                    self.Reset_Inventory_Slot(sending_inventory)
-                    return False
-            except TypeError as e:
-                print(f"Receiving inventory not a weapon inventory: {e}")
-        return True
-    
-    def Check_Two_Handed_Left_Hand(self, inventory_slot):
-        if self.entity.type == 'player':
-            self.entity.Set_Inventory_Interaction(20)
-
-        if not inventory_slot.inventory_type:
-            return True
-        if not 'two' in self.weapon_class:
-            return True
-        if 'left' in inventory_slot.inventory_type:
-            return True
-        else:
-            return False
-    
-    
-    def Check_For_Two_Handed_In_Weapon_Inventory(self, inventory_slot, sending_inventory, receiving_inventory):
-        # Check for weapon inventory
-        
-        if not inventory_slot.inventory_type:
-            return True
-        
-        for weapon_inventory_slot in receiving_inventory:
-            if not weapon_inventory_slot.item:
-                continue
-            # Check to ensure that item is a weapon
-            if not weapon_inventory_slot.item.category == 'weapon':
-                continue
+    #         break
             
-            # Check if there is already a two handed weapon in the inventory
-            if 'two' in weapon_inventory_slot.item.weapon_class:
-                self.Reset_Inventory_Slot(sending_inventory)
-
-                return False
         
-        return True
+    #     if not recieving_inventory_slot:
+    #         self.Reset_Inventory_Slot(sending_inventory)
+    #         return False
+    #     # Check if we can send the item to the new inventroy slot
+    #     if self.Send_To_Inventory(recieving_inventory_slot, sending_inventory, receiving_inventory):
+    #         return True
+    #     else:
+    #         self.Reset_Inventory_Slot(sending_inventory)
+    #         return False
+
+    # def Check_Legal_Move(self, inventory_slot, sending_inventory, receiving_inventory):
+    #     if not inventory_slot:
+    #         return False
+        
+    #     if not self.Check_Two_Handed(inventory_slot, sending_inventory, receiving_inventory):
+    #         return False
+    
+    #     if not self.Check_For_Two_Handed_In_Weapon_Inventory(inventory_slot, sending_inventory, receiving_inventory):
+    #         return False
+        
+    #     if not self.Check_Two_Handed_Left_Hand(inventory_slot):
+    #         return False
+        
+    #     if not self.Bow_Check(inventory_slot):
+    #         return False
+        
+    #     if not self.Arrow_Check(inventory_slot):
+    #         return False
+        
+    #     return True
+
+
+    # # Attempt to move the item to the receiving inventory slot by double clicking
+    # def Send_To_Inventory(self, inventory_slot, sending_inventory, receiving_inventory):
+       
+    #     if not self.Check_Legal_Move(inventory_slot, sending_inventory, receiving_inventory):
+    #         return False
+
+    #     # Move the item
+    #     move_successful = receiving_inventory.Move_Item(self, inventory_slot)
+    #     # If the move was successful, remove it from the sending inventory
+    #     if move_successful:
+    #         self.Reset_Weapon()
+    #         # Remove item from old inventory and save the inventory type
+    #         inventory_type_holder = self.inventory_type
+
+    #         sending_inventory.Remove_Item(self, True)
+    #         # send weapon into weapon inventory, checked by seeing if it has an inventory_type
+    #         if self.inventory_type:
+
+    #             # If weapon is already equipped in the other hand, remove it before adding to new hand
+    #             if self.equipped:
+    #                 self.game.player.Remove_Active_Weapon(inventory_type_holder)
+    #             self.Equip()
+    #         else: # Drag weapon back into item inventory
+    #             self.Set_Equip(False)
+    #             self.game.player.Remove_Active_Weapon(inventory_type_holder)
+
+    #         return True
+        
+        
+    #     return False
+    
+    # def Reset_Weapon(self):
+    #     self.rotate = 0
+    #     self.attacking = 0
+    #     self.sub_type = self.type
+    #     self.animation = 0
+
+    # # Add weapons from a sending inventory to a receiving inventory using drag
+    # # Example: weapon_inventory (sending) -> inventory (receiving) 
+    # def Move_To_Other_Inventory(self, sending_inventory, receiving_inventory, offset = (0,0)):
+    #     if self.game.mouse.left_click:
+    #         return
+    #     # Check for collision with new ivnentory slot
+    #     for receiving_inventory_slot in receiving_inventory:
+    #         if receiving_inventory_slot.rect().colliderect(self.game.mouse.rect_pos(offset)):
+    #             if self.Send_To_Inventory(receiving_inventory_slot, sending_inventory, receiving_inventory):
+    #                 for inventory_slot in sending_inventory.inventory:
+    #                     if not inventory_slot.active: # Reset active state after dragging
+    #                         continue
+    #                     inventory_slot.Set_Active(False)
+    #                     break
+    #                 return True             
+    #     return False
+
+    # # Check if the weapon can be moved to the weapon inventory
+    # def Move_Inventory_Check(self, offset = (0,0)):
+    #     if not self.picked_up:
+    #         active_inventory = self.game.weapon_inventory.active_inventory
+    #         weapon_inventory = self.game.weapon_inventory.inventories[active_inventory]
+    #         if self.equipped: # Move to normal inventory
+    #             if self.Move_To_Other_Inventory(weapon_inventory, self.game.item_inventory, offset):
+    #                 self.Set_Equip(False)
+    #                 self.game.player.Remove_Active_Weapon(self.inventory_type)
+    #                 return True
+                
+                
+    #         else: # Move to weapon inventory
+    #             if self.Move_To_Other_Inventory(self.game.item_inventory, weapon_inventory, offset):
+    #                 self.Equip()
+    #                 return True
+                
+    #     return False
+    
+    # # Equip the weapon
+    def Equip(self):
+        self.Set_Equip(True)
+        self.Set_Entity(self.game.player)
+        self.game.player.Set_Active_Weapon(self)
+
+
+
+
+    # # Check for out of bounds, return true if valid, else false
+    # def Move_Legal(self, mouse_pos, player_pos, tilemap, offset = (0,0)):
+
+    #     # Check if the weapon can be moved to the weapon inventory
+    #     if self.Move_Inventory_Check(offset):
+    #         self.picked_up = True
+    #         self.move_inventory_slot = True
+    #         return False
+    #     if super().Move_Legal(mouse_pos, player_pos, tilemap, offset):
+    #         return True
+    #     else:
+    #         return False
+        
+    # def Check_Inventory_Type(self, type):
+    #     legal_inventories = ['left_hand', 'right_hand', 'bow', 'arrow']
+    #     if type not in legal_inventories:
+    #         return False
+        
+    #     self.inventory_type = type
+    #     return True
+
+    # def Update_Player_Hand(self, prev_hand):
+    #     # Check if the weapon has changed hands
+    #     if self.inventory_type != prev_hand:
+    #         self.Set_Equip(True)
+    #         if prev_hand == 'left_hand':
+    #             self.game.player.Remove_Active_Weapon(prev_hand)
+    #             self.game.player.Set_Active_Weapon(self, self.inventory_type)
+    #         elif prev_hand == 'right_hand':
+    #             self.game.player.Remove_Active_Weapon(prev_hand)
+    #             self.game.player.Set_Active_Weapon(self, self.inventory_type)
+    #         elif 'bow' in prev_hand:
+    #             self.game.player.Remove_Active_Weapon(prev_hand)
+    #             self.game.player.Set_Active_Weapon(self, self.inventory_type)
+
+
+    # def Place_Down(self):
+    #     if not super().Place_Down():
+    #         return False
+    #     self.entity = None
+    #     self.in_inventory = False
+    #     self.picked_up = False
+    #     if self.equipped:
+    #         self.game.player.Remove_Active_Weapon(self.inventory_type)
+    #         self.Set_Equip(False)
+    #     return True
+
+    # def Set_In_Inventory(self, state):
+    #     self.in_inventory = state
+
+    # def Bow_Check(self, inventory_slot):
+    #     if not inventory_slot.inventory_type:
+    #         return True
+    #     if 'bow' in inventory_slot.inventory_type:
+    #         if 'bow' in self.type:
+    #             return True
+    #         else:
+    #             return False
+    #     return True
+        
+        
+    # def Arrow_Check(self, inventory_slot):
+
+    #     if not inventory_slot.inventory_type:
+    #         return True
+    #     if 'arrow' in inventory_slot.inventory_type:
+    #         if 'arrow' in self.weapon_class:
+    #             return True
+    #         else:
+    #             return False
+    #     return True
+
+    # def Check_Two_Handed(self, inventory_slot, sending_inventory, receiving_inventory):
+    #     if not 'two' in self.weapon_class:
+    #         return True
+    #     if inventory_slot.inventory_type:
+    #         # Try to find the inventory_slot, only the weapon inventory has this property
+    #         try:
+    #             if receiving_inventory.Find_Inventory_Slot(inventory_slot):
+    #                 # Find the original position of the item in the inventory
+    #                 self.Reset_Inventory_Slot(sending_inventory)
+    #                 return False
+    #         except TypeError as e:
+    #             print(f"Receiving inventory not a weapon inventory: {e}")
+    #     return True
+    
+    # def Check_Two_Handed_Left_Hand(self, inventory_slot):
+    #     if self.entity.type == 'player':
+    #         self.entity.Set_Inventory_Interaction(20)
+
+    #     if not inventory_slot.inventory_type:
+    #         return True
+    #     if not 'two' in self.weapon_class:
+    #         return True
+    #     if 'left' in inventory_slot.inventory_type:
+    #         return True
+    #     else:
+    #         return False
+    
+    
+    # def Check_For_Two_Handed_In_Weapon_Inventory(self, inventory_slot, sending_inventory, receiving_inventory):
+    #     # Check for weapon inventory
+        
+    #     if not inventory_slot.inventory_type:
+    #         return True
+        
+    #     for weapon_inventory_slot in receiving_inventory:
+    #         if not weapon_inventory_slot.item:
+    #             continue
+    #         # Check to ensure that item is a weapon
+    #         if not weapon_inventory_slot.item.category == 'weapon':
+    #             continue
+            
+    #         # Check if there is already a two handed weapon in the inventory
+    #         if 'two' in weapon_inventory_slot.item.weapon_class:
+    #             self.Reset_Inventory_Slot(sending_inventory)
+
+    #             return False
+        
+    #     return True
           
     
 
-    def Reset_Inventory_Slot(self, inventory_slot):
-        original_inventory_slot = inventory_slot.Find_Item_In_Inventory(self)
-        # Reset it back to not active if found
-        if original_inventory_slot:
-            original_inventory_slot.Set_Active(False)
-        return
-    ####################################################### 
+    # def Reset_Inventory_Slot(self, inventory_slot):
+    #     original_inventory_slot = inventory_slot.Find_Item_In_Inventory(self)
+    #     # Reset it back to not active if found
+    #     if original_inventory_slot:
+    #         original_inventory_slot.Set_Active(False)
+    #     return
+    # #######################################################
