+++ scripts/game/logic_update.py
@@ -14,9 +14,7 @@ class Logic_Update():
             self.game.entities_render.Update()
 
 
-            self.game.item_inventory.Update(self.game.render_scroll)
-            self.game.weapon_inventory.Update(self.game.render_scroll)
-            self.game.rune_inventory.Update(self.game.render_scroll)
+            self.game.inventory.Update(self.game.render_scroll)
             self.game.souls_interface.Update()
             self.game.rune_handler.Update(self.game.render_scroll)
             self.game.player.Update(self.game.tilemap, (self.game.movement[1] - self.game.movement[0], self.game.movement[3] - self.game.movement[2]), self.game.render_scroll)
