+++ scripts/game/renderer.py
@@ -23,9 +23,7 @@ class Renderer():
         self.game.rune_handler.Render_Animation(surf, render_scroll)
         
 
-        self.game.item_inventory.Render(surf)
-        self.game.weapon_inventory.Render(surf)
-        self.game.rune_inventory.Render(surf)
+        self.game.inventory.Render(surf)
 
         self.game.text_box_handler.Render(surf, render_scroll)
         self.game.player.effects.Render_Effects_Symbols(surf)
